+++ src/pyochain/__init__.py
@@ -5,7 +5,7 @@ from ._set import Set, SetMut
 from ._sliceview import SliceView
 from ._tools import Iter, Peekable  # pyright: ignore[reportMissingModuleSource]
 from ._vec import Vec
-from ._views import PyoItemsView, PyoKeysView, PyoValuesView
+from .abc import PyoItemsView, PyoKeysView, PyoValuesView
 from .rs import (
     NONE,
     Err,